+++ src/database/migrations/2018_01_07_153340_create_contract_items_table.php
@@ -16,7 +16,7 @@ class CreateContractItemsTable extends Migration
 
         Schema::create('contract_items', function (Blueprint $table) {
 
-            $table->integer('contract_id');
+            $table->bigInteger('contract_id');
             $table->bigInteger('record_id');
             $table->integer('type_id');
             $table->integer('quantity');
